fileupload

+++ MovieApp/src/main/java/com/test/mapper/BoardMapper.java
@@ -9,7 +9,7 @@ public interface BoardMapper {
 	
 	public List<BoardVO> getList();
 	
-	public BoardVO get(Long bno);
+	public BoardVO get(int bno);
 	
 	public void insert (BoardVO board);
